rename: side-bar-icons→menu-item-icons
