+++ components/side-bar/menu-item-icons/AuditIcon.tsx
