+++ components/side-bar/menu-item-icons/HomeIcon.tsx
