+++ components/side-bar/menu-item-icons/MenuIcon.tsx
