+++ components/side-bar/menu-item-icons/SalesIcon.tsx
