+++ components/side-bar/menu-item-icons/SettingsIcon.tsx
