+++ components/side-bar/menu-item-icons/ShiftIcon.tsx
