+++ components/side-bar/menu-item-icons/UserIcon.tsx
